Get rid of implicit context
