+++ src/index.ts
@@ -1,4 +1,4 @@
 export * from "./api";
-export { HandlerInfo, handlerLinks, extractHandlerInfo, inHandlerContext } from "./handler";
+export * from "./handler";
 export * from "./operation";
 export * from "./serializer";
