+++ src/operation.ts
@@ -1,91 +1,3 @@
-import { OperationInfo, Link, HandlerError } from "./api";
-import { LazyValue } from "./serializer";
-
-/**
- * Options for the {@link OperationHandler["start"]} method.
- */
-export interface StartOperationOptions {
-  /** Signaled when the current request is canceled. */
-  abortSignal: AbortSignal;
-  /** Request headers. */
-  headers: Record<string, string>;
-  /**
-   * Callbacks are used to deliver completion of async operations.
-   * This value may optionally be set by the client and should be called by a handler upon completion if the started
-   * operation is async.
-   */
-  callbackURL?: string;
-  /**
-   * Optional header fields set by a client that are required to be attached to the callback request when an
-   * asynchronous operation completes.
-   */
-  callbackHeaders?: Record<string, string>;
-  /**
-   * Request ID that may be used by the server handler to dedupe a start request.
-   * By default a v4 UUID will be generated by the client.
-   */
-  requestId?: string;
-  /**
-   * Links contain arbitrary caller information. Handlers may use these links as metadata on resources associated with
-   * and operation.
-   */
-  links: Link[];
-}
-
-/**
- * Options for the {@link OperationHandler["getInfo"]} method.
- */
-export interface GetOperationInfoOptions {
-  /** Signaled when the current request is canceled. */
-  abortSignal: AbortSignal;
-  /** Request headers. */
-  headers: Record<string, string>;
-}
-
-/**
- * Options for the {@link OperationHandler["getResult"]} method.
- */
-export interface GetOperationResultOptions {
-  /** Signaled when the current request is canceled. */
-  abortSignal: AbortSignal;
-  /** Request headers. */
-  headers: Record<string, string>;
-  /**
-   * If specified and non-zero, reflects the duration the caller has indicated that it wants to wait for operation
-   * completion, turning the request into a long poll.
-   */
-  wait: number;
-}
-
-/**
- * Options for the {@link OperationHandler["cancel"]} method.
- */
-export interface CancelOperationOptions {
-  /** Signaled when the current request is canceled. */
-  abortSignal: AbortSignal;
-  /** Request headers. */
-  headers: Record<string, string>;
-}
-
-/** A result that indicates that an operation completed successfully. */
-export interface HandlerStartOperationResultSync<T> {
-  value: T;
-}
-
-/** A result that indicates that an operation has been accepted and will complete asynchronously. */
-export interface HandlerStartOperationResultAsync {
-  /**
-   * A token to identify the operation in followup handler methods such as {@link OperationHandler["getResult"]} and
-   * {@link OperationHandler["cancel"]}.
-   */
-  token: string;
-}
-
-/**
- * The return type from the {@link OperationHandler["start"]}. May be either a synchronous or asynchornous.
- */
-export type HandlerStartOperationResult<T> = HandlerStartOperationResultSync<T> | HandlerStartOperationResultAsync;
-
 declare const inputBrand: unique symbol;
 declare const outputBrand: unique symbol;
 
@@ -97,61 +9,6 @@ export interface Operation<I, O> {
   [inputBrand]: I;
   [outputBrand]: O;
 }
-
-/**
- * A handler for an operation.
- */
-export interface OperationHandler<I, O> {
-  /**
-   * Handles requests for starting an operation.
-   *
-   * Return {@link HandlerStartOperationResultSync} to respond successfully - inline, or
-   * {@link HandlerStartOperationResultAsync} to indicate that an asynchronous operation was started. Throw a
-   * {@link OperationError} to indicate that an operation completed as failed or canceled.
-   */
-  start(input: I, options: StartOperationOptions): Promise<HandlerStartOperationResult<O>>;
-
-  /**
-   * Handles requests to get the result of an asynchronous operation. Return non error result to respond successfully -
-   * inline, or error with {@link OperationStillRunningError} to indicate that an asynchronous operation is still
-   * running.
-   *
-   * Throw an {@link OperationError} to indicate that an operation completed as failed or canceled.
-   *
-   * When {@link GetOperationResultOptions["wait"]} is greater than zero, this request should be treated as a long poll.
-   * Note that the specified wait duration may be longer than the configured client or server side request timeout, and
-   * should be handled separately.
-   *
-   * It is the implementor's responsiblity to respect the client's wait duration and return in a timely fashion, leaving
-   * enough time for the request to complete and the response to be sent back.
-   *
-   * @experimental
-   */
-  getResult(token: string, options: GetOperationResultOptions): Promise<O>;
-
-  /**
-   * GetInfo handles requests to get information about an asynchronous operation.
-   *
-   * @experimental
-   */
-  getInfo(token: string, options: GetOperationInfoOptions): Promise<OperationInfo>;
-
-  /**
-   * Handles requests to cancel an asynchronous operation.
-   *
-   * Cancelation in Nexus is:
-   * 1. asynchronous - returning from this method only ensures that cancelation is delivered, it may later be
-   * ignored by the underlying operation implemention.
-   * 2. idempotent - implementors should ignore duplicate cancelations for the same operation.
-   */
-  cancel(token: string, options: CancelOperationOptions): Promise<void>;
-}
-
-/**
- * A shortcut for defining an operation handler that only implements the {@link OperationHandler["start"]} method and
- * always returns a {@link HandlerStartOperationResultSync}.
- */
-export type SyncOperationHandler<I, O> = (input: I, options: StartOperationOptions) => Promise<O>;
 
 /**
  * A named collection of operation handlers.
@@ -168,19 +25,6 @@ export type PartialOperationMap = Record<string, PartialOperation<any, any>>;
  */
 export type OperationMapFromPartial<T extends PartialOperationMap> = {
   [K in keyof T & string]: T[K] extends PartialOperation<infer I, infer O> ? Operation<I, O> : never;
-};
-
-/**
- * A type that defines a handler for a given operation.
- */
-export type OperationHandlerFor<T> =
-  T extends Operation<infer I, infer O> ? OperationHandler<I, O> | SyncOperationHandler<I, O> : never;
-
-/**
- * A type that defines a collection of handlers for a given collection of operation interfaces.
- */
-export type ServiceHandlerFor<T extends OperationMap = OperationMap> = {
-  [K in keyof T & string]: OperationHandlerFor<T[K]>;
 };
 
 /**
@@ -265,148 +109,3 @@ export type OperationInput<T> = T extends Operation<infer I, any> ? I : any;
  * A type that extracts the output type from an operation in a service.
  */
 export type OperationOutput<T> = T extends Operation<any, infer O> ? O : any;
-
-/**
- * A {@link Service} that includes a collection of handlers for its operations.
- */
-export interface ServiceHandler<T extends OperationMap = OperationMap> extends Service<T> {
-  handlers: ServiceHandlerFor<T>;
-}
-
-/**
- * Constructs a service handler for a given service contract.
- */
-export function serviceHandler<T extends OperationMap>(
-  service: Service<T>,
-  handlers: ServiceHandlerFor<T>,
-): ServiceHandler<T> {
-  const ops = new Set<string>();
-
-  for (const [k, op] of Object.entries(service.operations)) {
-    if (!op.name) {
-      throw new TypeError(`Tried to register an operation with no name for service ${service.name} with key ${k}`);
-    }
-    if (ops.has(op.name)) {
-      throw new TypeError(`Operation with name ${op.name} already registered for service ${service.name}`);
-    }
-    const handler = handlers[k];
-    if (!handler) {
-      throw new TypeError(`No handler registered for ${k} on service ${service.name}`);
-    }
-    ops.add(op.name);
-  }
-
-  return {
-    ...service,
-    handlers,
-  };
-}
-
-/**
- * A collection of service handlers that dispatches requests to the registered service and operation handler.
- */
-export class ServiceRegistry {
-  private services = new Map<string, Map<string, OperationHandler<any, any> | SyncOperationHandler<any, any>>>();
-
-  constructor(services: ServiceHandler[]) {
-    for (const s of services) {
-      if (!s.name) {
-        throw new TypeError("Tried to register a Nexus service with no name");
-      }
-      if (this.services.has(s.name)) {
-        throw new TypeError(`Duplicate registration of nexus service ${s.name}`);
-      }
-      const ops = new Map<string, OperationHandler<any, any> | SyncOperationHandler<any, any>>();
-      for (const [k, op] of Object.entries(s.operations)) {
-        if (!op.name) {
-          throw new TypeError(`Tried to register an operation with no name for service ${s.name} with key ${k}`);
-        }
-        if (ops.has(op.name)) {
-          throw new TypeError(`Operation with name ${op.name} already registered for service ${s.name}`);
-        }
-        const handler = s.handlers[k];
-        if (!handler) {
-          throw new TypeError(`No handler registered for ${k} on service ${s.name}`);
-        }
-        ops.set(op.name, handler);
-      }
-      this.services.set(s.name, ops);
-    }
-  }
-
-  private getHandler(service: string, operation: string): OperationHandler<any, any> | SyncOperationHandler<any, any> {
-    const serviceHandler = this.services.get(service);
-    if (serviceHandler == null) {
-      throw new HandlerError({
-        type: "NOT_FOUND",
-        message: `Service handler not registered for service ${service}`,
-      });
-    }
-    const operationHandler = serviceHandler.get(operation);
-    if (operationHandler == null) {
-      throw new HandlerError({
-        type: "NOT_FOUND",
-        message: `Operation handler not registered for operation ${operation} in service ${service}`,
-      });
-    }
-    return operationHandler;
-  }
-
-  async start(
-    service: string,
-    operation: string,
-    lv: LazyValue,
-    options: StartOperationOptions,
-  ): Promise<HandlerStartOperationResult<any>> {
-    const handler = this.getHandler(service, operation);
-    const input = await lv.consume<any>();
-    if (typeof handler === "function") {
-      const value = await handler(input, options);
-      return { value };
-    }
-    return await handler.start(input, options);
-  }
-
-  async getResult(
-    service: string,
-    operation: string,
-    token: string,
-    options: GetOperationResultOptions,
-  ): Promise<LazyValue> {
-    const handler = this.getHandler(service, operation);
-    if (typeof handler === "function") {
-      throw new HandlerError({
-        type: "NOT_IMPLEMENTED",
-        message: "Not implemented",
-      });
-    }
-    return await handler.getResult(token, options);
-  }
-
-  async getInfo(
-    service: string,
-    operation: string,
-    token: string,
-    options: GetOperationInfoOptions,
-  ): Promise<OperationInfo> {
-    const handler = this.getHandler(service, operation);
-    if (typeof handler === "function") {
-      throw new HandlerError({
-        type: "NOT_IMPLEMENTED",
-        message: "Not implemented",
-      });
-    }
-    return await handler.getInfo(token, options);
-  }
-
-  async cancel(service: string, operation: string, token: string, options: CancelOperationOptions): Promise<void> {
-    const handler = this.getHandler(service, operation);
-    if (typeof handler === "function") {
-      throw new HandlerError({
-        type: "NOT_IMPLEMENTED",
-        message: "Not implemented",
-      });
-    }
-    return await handler.cancel(token, options);
-  }
-}
